+++ BackEnd/queries/usersQueries.js
@@ -6,17 +6,10 @@ const pool = new Pool({
     user: 'postgres',
     host: 'localhost',
     database: 'nutribuddy',
-    password: '',
+    password: 'password',
     port: 5432,
 })
 
-pool.connect((error) => {
-    if (error) {
-        console.error("error connecting to db", error);
-    }else {
-        console.log("connected to db");
-    }
-});
 //////////////////////////////////////////////////////////////////////////////////
 // Users Table Query
 const getAllUsers = (request, response) => {
@@ -182,200 +175,11 @@ const addUser = (request, response) => {
     }
 };
 
-//////////////////////////////////////////////////////////////////////////////////
-// Meals Table Query
-const getUserMealsByUserId = (request, response) => {
-    const id = parseInt(request.params.id);
-    pool.query(`SELECT * FROM meal WHERE user_id = $1`, [id], (error, results) => {
-        if (error) {
-            throw error;
-        }
-        response.status(200).json(results.rows);
-    });
-};
-
-const getUsersMealsByPastDate = (request, response) => {
-    const id = parseInt(request.params.id);
-    const inputDate = request.params.date;
-    pool.query('SELECT * FROM meal WHERE user_id = $1 AND meal_time < $2', [id, inputDate], (error, results) => {
-        if (error) {
-            throw error;
-        }
-        response.status(200).json(results.rows);
-    });
-};
-
-const getUsersMealsByFutureDate = (request, response) => {
-    const id = parseInt(request.params.id);
-    const inputDate = request.params.date;
-    pool.query('SELECT * FROM meal WHERE user_id = $1 AND meal_time >= $2', [id, inputDate], (error, results) => {
-        if (error) {
-            throw error;
-        }
-        response.status(200).json(results.rows);
-    });
-};
-
-const getUsersMealsByMealType = (request, response) => {
-    const id = parseInt(request.params.id);
-    const mealType = request.params.type;
-    pool.query('SELECT * FROM meal WHERE user_id = $1 AND meal_type = $2', [id, mealType], (error, results) => {
-        if (error) {
-            throw error;
-        }
-        response.status(200).json(results.rows);
-    });
-};
-
-const getUsersMealsByCalories = (request, response) => {
-    const id = parseInt(request.params.id);
-    const calories = request.params.calories;
-    pool.query('SELECT * FROM meal WHERE user_id = $1 AND meal_calories <= $2', [id, calories], (error, results) => {
-        if (error) {
-            throw error;
-        }
-        response.status(200).json(results.rows);
-    });
-};
-
-const addUserMeal = (request, response) => {
-    try {
-        const {
-            recipe_id,
-            user_id,
-            meal_name,
-            meal_time,
-            meal_type,
-            meal_calories
-        } = request.body;
-        pool.query(
-            `INSERT INTO meal (recipe_id, user_id, meal_name, meal_time, meal_type, meal_calories)
-             VALUES ($1, $2, $3, $4, $5, $6) RETURNING *`,
-            [recipe_id, user_id, meal_name, meal_time, meal_type, meal_calories],
-            (error, results) => {
-                if (error) {
-                    console.log(error, '<--- error here')
-                    throw error;
-                }
-                response.status(201).json(results.rows)
-            }
-        );
-    } catch (error) {
-        throw error;
-    }
-};
-
-const deleteUserMealById = (request, response) => {
-    const id = parseInt(request.params.id);
-    pool.query(`DELETE FROM meal WHERE id = ${id}`, (error, results) => {
-        if(error) {
-            throw error;
-        }
-        response.status(200).json(results.rows);
-    });
-};
-//////////////////////////////////////////////////////////////////////////////////
-// Ingredients Table Query
-const getIngredientsByMealId = (request, response) => {
-    const id = request.params.id;
-    pool.query(`SELECT * FROM ingredients WHERE recipe_id = $1`, [id], (error, results) => {
-        if(error) {
-            throw error;
-        }
-        response.status(200).json(results.rows);
-    });
-};
-
-const addIngredients = (request, response) => {
-    try {
-        const {
-            recipe_id,
-            ingredient_name,
-            quantity,
-            unit_type,
-            price_per_unit
-        } = request.body;
-        pool.query(
-            `INSERT INTO ingredients (recipe_id, ingredient_name, quantity, unit_type, price_per_unit)
-             VALUES ($1, $2, $3, $4, $5) RETURNING *`,
-            [recipe_id, ingredient_name, quantity, unit_type, price_per_unit],
-            (error, results) => {
-                if (error) {
-                    console.log(error, '<--- error here')
-                    throw error;
-                }
-                response.status(201).json(results.rows)
-            }
-        );
-    } catch (error) {
-        throw error;
-    }
-};
-//////////////////////////////////////////////////////////////////////////////////
-// Favorite Table Query
-const getFavMealsById = (request, response) => {
-    const id = parseInt(request.params.id);
-    pool.query(`SELECT recipe_name FROM favorite_meals WHERE user_id = ${id}`, (error, results) => {
-        if (error) {
-            throw error;
-        }
-        response.status(200).send(results.rows)
-    });
-};
-
-const addFavMeal = (request, response) => {
-    try {
-        const {
-            recipe_name,
-            recipe_id,
-            user_id
-        } = request.body;
-        pool.query(
-            `INSERT INTO favorite_meals (recipe_name, recipe_id, user_id) VALUES ($1, $2, $3) RETURNING *`,
-            [recipe_name, recipe_id, user_id],
-            (error, results) => {
-                if (error) {
-                    console.log(error, '<--- error here')
-                    throw error;
-                }
-                response.status(201).json(results.rows)
-            }
-        );
-    } catch (error) {
-        throw error;
-    }
-};
-
-const deleteFavMeal = (request, response) => {
-    const id = parseInt(request.params.id);
-    pool.query(`DELETE FROM favorite_meals WHERE id = ${id}`, (error, results) => {
-        if(error) {
-            throw error;
-        }
-        response.status(200).json(results.rows);
-    });
-};
-
 module.exports = {
     getAllUsers,
     getUserById,
     deleteUserById,
     updateUserPersonalById,
     updateUserAllergiesById,
-    addUser,
-
-    getUserMealsByUserId,
-    getUsersMealsByPastDate,
-    getUsersMealsByFutureDate,
-    getUsersMealsByMealType,
-    getUsersMealsByCalories,
-    addUserMeal,
-    deleteUserMealById,
-
-    getIngredientsByMealId,
-    addIngredients,
-
-    getFavMealsById,
-    addFavMeal,
-    deleteFavMeal
+    addUser
 }
